Fixes to separation of nlp and tabular
